Update Paella and add support for fallback fullscreen mode on iOS
Fixes #1070 -> well, there is still one warning remaining, but that
needs to be fixed inside Paella and is more complicated to get rid of.
See https://github.com/polimediaupv/paella-core/issues/354

Fixes #990 -> well, this doesn't add volume controls back but the
reasoning for hiding those is somewhat sane, so that part is #wontfix.

The fullscreen fallback mode for Safari was suggested here:
https://github.com/polimediaupv/paella-core/pull/355
But it was moved into the plugin by the Paella devs.