Future proof harvest sync code for new items
Before this commit, when the harvest API would have returned an item
with a `kind` not known to Tobira, harvest would fail. It is expected
that a new kind, namely playlists, will be added with Opencast 16. We
will of course extend the harvest code to properly store playlists then.
This commit just makes sure that Tobira can also deal with these future
APIs without breaking.