Update crypto-related crates
This was a fairly annoying update as it contains many breaking changes.
And `ring` is annoying as usual by not providing a changelog. Now that
rustls supports other crypto providers, maybe we can get rid of ring
at some point.

To be honest, I have not tested all code paths. I think it should work,
but we should probably try loading a certificate and using TLS with the
DB before the next release... or ideally before this PR is merged.