Make big chunk of `realm` Playwright tests work again
This was rewritten with the new fixed data and isolation model in mind.
I decided to leave danger zone operations for now, as they should
probably live in a new test.

I moved around and changed a few utility function. I don't think
`realmSetup` will be needed anymore for the blocks function, so it is
removed completely. I renamed `blocks.spec.ts` to disable it completely
without causing a big diff.