Add special roles for global page admins & moderators
This was previously what moderators could do, but this changed by
introducing the granular realm permissions. But there might still be
use for these users who can just edit all non-user realms.

I haven't added these roles to any of our dummy users, as we want to
somewhat protect the test deployment now. This means not giving public
access to edit the home page.

This commit also adds an entry in the ACL UI for page admins, with the
same reasoning as with normal admins: it might be weird not seeing any
entries applying to you and still apparently having access. Though the
user-facing strings for this entry... are not great.