Removes "Tobira moderators" and adds "Tobira admins"
See #1133 for context. The moderator role got weird over time and it
was unclear what it is supposed to be able to do. In the previous
commits specific roles for specific actions were added. But it's useful
to just give all of those privileges to some users, but without giving
them the nuclear option ROLE_ADMIN. Unlike "Tobira moderator", it's
pretty clear what a "Tobira admin" can do: everything inside Tobira.