Improve some log statements by using fields
That's one of the powerful things about `tracing`: we don't have to
interpolate all data somehow into our string message, but can just
specify them directly. Yes, the textual formatting might be a bit worse
in some cases, but the ability to get that raw data directly is worth
it. It's also a bit nicer because extra data is consistently shown, not
depending on how the programmer decides to interpolate.