Completely switch to `tracing` using own formatter & subscribers
In the previous commit, the log-tracing compatibility worked in one
direction: all tracing events were converted to log events, and our
log implementation emitted them. Now it's the other way around: all log
events get converted to tracing ones and we have a tracing subscriber
that emits those events.

This means that basically all of `logger.rs` was rewritten: the
filtering and formatting now works with `tracing` instead of `log`.
The formatting code is unfortunately pretty ugly, but I feel like this
is somewhat unavoidable for this task. Lots of branches and stuff.

I tried to keep the format very similar to the previous one, but of
course the base data is different: tracing events not only have a
message & metadata, but also fields. And those have to be formatted
somehow.