Let all users use the editor on test deployment & local environment
Instead of giving that privilege to only Sabine and Björk in the last
commit, I thought: why not use the role config at least once in our
test deployment. To kind of "test" that as well. And I don't see a
problem letting everyone use the editor (it's only allowed for videos
you have write access to anyway).