Adjust spacing of forwards and backwards button text
Since these are basically mirrored (not completely because
the number 10 is obv not), the x-offset does not necessarily
need to be the same. It's kinda hard to tell honestly, so
I tried around a bit with different values and think to have
ended up with an improved balancing. Though that is still somewhat
subjective of course.