"Improve" edit-button locating
This replaces the `nth()` locator with more specific ones.
Using `nth()` is apparently discouraged and indeed lead
to some flaky behaviour in the block editing tests.
The buttons are now located by using specific unique locators
for each kind of block, in combination with the xPath ("..")
locator to get to the parent. This does however make the
locators pretty ugly. Using ("..") once or maybe twice is fine
but in the case of the edit buttons, it needs to be chained up to 8
times which is just ridiculous. If there is a nicer way, I haven't
located it.

This also incidentally inlines all the block editing instead of
having one large function for it. I do maintain that the function
was nicer, but with these new locators it has gotten somewhat
pointless.