Improve edit-button locating
This replaces the `nth()` locator with more specific ones.
Using `nth()` is apparently discouraged and indeed lead
to some flaky behaviour in the block editing tests.
The buttons are now located by using specific unique locators.
Blocks containing only text content like title and text blocks
can be located by using a regex filter, while the video and series
blocks can be located by combining a filter that checks the
presence of specific text with a filter that checks for
the absence of other specific text (that is contained in another
block).
However, we do need to make sure that the test steps are
performed in a set order: If the text whose presence is checked
is removed in a test step, that step needs to be the very last one
of that specific block. Otherwise the locator will fail.

This also incidentally inlines all the block editing instead of
having one large function for it. I do maintain that the function
was nicer, but with these new locators it has gotten somewhat
pointless.