Update `reinda` and improve asset handling
This finally (!!) revamps our asset handling. We use reinda, which does
a lot for us, but it had some limitations before. Specifically, it
could not embed files by wildcard pattern and files could not be added
purely at runtime. This made frontend code splitting and including
custom fonts impossible. To improve this, reinda was almost completely
rewritten.

The most important changes from reinda and this commit:
- No more reinda templates: it has been replaced by the "modifier" API
  in reinda, which gives Tobira more control. This gets rid of the
  million manual variables in our index.html.
- The `embed!` macro only mentions files that are actually embedded into
  the executable. Logos for example are added at runtime, without a
  weird "path_override".
- The compression of embedded assets is now done with brotli, which
  should result in a smaller binary.
- The "logo fallback logic" is now performed in the frontend. This is
  more convenient for us and avoids duplicating the logo contents in
  memory.
- Fonts and Paella icons are embedded by wildcard now, so we don't have
  to list every darn file anymore. I am slightly uneasy as this means in
  the future, we might accidentally include unneeded files there.

I don't think reinda's API is "perfect" now. My main complain is the
duplication of paths/filenames. But we want to ship that at some point
and I think the API works well enough. The duplication here is unlikely
to result in bad bugs that are only noticed late.

NOTE: This does not restore full previous functionality yet! Two things
are missing:
- The font paths in `fonts.css` are not adjusted and broken.
- The frontend code map is not included yet.

Both of these will be fixed in the next commits.