Enable frontend bundle code splitting (Paella as own bundle)
This is the first step to using code splitting to reduce the loaded JS
size. This commit lays the ground work but only splits the bundle into
two files. But that's already a big win as it cuts the size of JS one
has to download to render the main page roughly in half. Even if there
is a player on the main page, the page already renders once the main
bundle is loaded, and the player shows a placeholder spinner.

In the future, we can create more bundles, e.g. moving everything
"manage" related to another bundle or even just letting webpack decide.