Use webpack-html-plugin to insert bundle path instead of backend
This is more straight forward. Before this commit, the setup was broken
as the "find main bundle" code in `assets.rs` was incorrect. It only
iterated over files embedded at compile time, which might be missing the
bundle. But more: if the frontend was changed, the hash and thus
filename of the main bundle changed as well, and the backend replacement
would be incorrect.

Letting webpack handle this part seems simpler.

This commit also deduplicates some strings in `assets.rs`.