Improve warning when harvest item cannot be deserialized
This is a band aid fix and not perfect. When I added the `Unknown`
variant to the enum, I wasn't aware that it was used when other variants
fail to deserialize even with `kind: "event"`. So the warning there
was misleading.

It's still not perfect because I ideally want to print more information
on the deseralization failure. And the `kind` strings are duplicated
now. I think in the future, the `HarvestResponse` might have items of
type `{ kind: String, #[serde(flatten)] body: serde_json::Value }` or
something like that. And then we deserialize it manually in a second
step. We will see.