Use Button from appkit
Moves the basic styled button, as well as its styling method, to appkit.
With this, we are now using components from appkit that use colors, so
we have to correctly configure out colors. Appkit has a lot more colors
than we have, but we only need to configure the ones actually used by
the button.