Add internal API functions to make http delete requests
The underlying function calls the external OC event API
to request the deletion of a single video and sets the
`tobira_deletion_timestamp` in the video's DB entry.

Unfortunately this API can't tell us if the deletion
succeeded, only that the retraction of the event was started.
That check and the following removal from the DB is done
by inspecting the previously mentioned timestamp (implemented
in the next commit).