Adjust "My Videos" page to show pending/failed deletions
"My Videos" will now indicate if a video has been deleted in
Tobira, but not (yet) in Opencast. After the request has been
sent, the deleted video will be marked as "delete pending".
When the event hasn't been removed in sync after a minute,
the deletion in Opencast has probably failed, so it gets marked
accordingly.