Add `createRealmLineage` mutation to API
This is supposed to make mounting series easier from external scripts.
The existing `mountSeries` API is (a) doing a lot of different things
and (b) requires the caller to distinguish between what realms already
exist and which still need to be created. This new API is idempotent and
more convenient. With this, it would be possible to simplify
`mountSeries`, but we can't simply break that API since the Admin UI
is using it. So we keep it for now. We might add an alternative API in
the future.