Fix overflowing videos
The video container width is set to `unset` when there
is enough space for the control elements. I know this
had a good-ish reason, I just don't remember it.
On that note, I should start writing better commit messages,
or at least try to explain things like that.

So anyway, this unset width can lead to situations were
videos with weird aspect ratios like 15/4 or sth would
overflow on smaller (but larger than mobile) screens.
To fix that without removing or complicating the `unset`
rule, this commit introduced a `maxWidth` of 100%.
BUT, that also needs to be overwritten on mobile screens
to preserve current sizing behaviour on these.