Add function for fetching text assets from OC (`sync texts fetch`)
This will be part of the worker and is able to deal with a variety of
error cases. Figuring all this out took quite some time. I decided now
that ignoring assets for which Opencast returns something unexpected is
fine most of the time. Admins will be able to easily requeue these
failed events.

This can also deal with network errors or similar indications that OC
is not available at the moment, using an exponential backoff then.