Optimize text search index
The main change is that texts with the same span are concatenated to
only be one entry in the index. This doesn't reduce the size of the
`texts` field in Meili, but that of the timespan index. This
optimization is mostly there for slide texts, not for captions.

But this commit also moves the build process into `FromSql` to avoid
a bunch of useless allocations. Ideally one would also avoid all the
intermediate `String` allocations, but that's not easily possible right
now.