Add playlist blocks
This involves several frontend components which are
related to the creation and editing of playlist blocks.
It features very little original code as these functions
and UI elements are already established for other block
types. Though a few of the adjustments necessary to make
everything work weren't super trivial either.