Change id type of playlist_entry in db
This was called `opencast_id` before, which was a little
misleading since it's not the kind of `opencast_id` used
in other places which is typically (but not necessariliy)
a uuid but can be any string.
The id that is now renamed to `entry_id` however is always a number.