Make `searchAll*` endpoints sort items by date when query is empty
This is useful for adding video/series/playlist blocks, as users are
very likely to include an item that they recently created/uploaded. Once
a search query is entered, the ordering reverts back to normal:
relevancy.

Solves #1013

This changes the search index, but the index version was already bumped
a few commits ago and there was no release in between.