Improve `textMatches` in search API by joining multiple matches
This rewrites the logic that creates the `textMatches` array for the
search API. Before, one Meili match was emitted as one text match, but
this had several problems. Most importantly, with two words in the
query, if those words would appear in a text right next to one another,
Meili would still generate two matches. They would have the same
timespan and Tobira would just show two divs on top of each other, only
one of which would be visible.

Now, for each individual text, we join all matches (with a limit) and
return only one `TextMatch`, but potentially with multiple highlight
ranges.