Add migration and API for new preview_roles
This is in preparation for upcoming changes regarding
event specific authentication.
Users with a preview role but without read access will
only be allowed to view text metadata like title and
description of events corresponding to that role.
Specifically they will not be allowed to watch the video
or access "revealing" information like thumbnail, slide
segments, captions or track data.
This will be implemented in the following commit.