Move some event data to `authorizedData` field
Authorized date is comprised of track information, segments,
caption texts and thumbnails.
This will be used to distinguish and handle any event data
that is not viewable and shouldn't be exposed by the api
when a user only has a `preview` role.

The presence or absence of this data can be used to determine
whether a user has a certain authorization level - this can
be useful for certain frontend checks.

This data will also get a credentials check which is to be performed
on the backend side in a later commit.
With that, it can be exposed when a user successfully authenticates.