Apply UI changes related to new preview roles and authorized data
Some video pages will now show a placeholder and note
instead of the player, while these videos in series blocks
and search results will have a "special" preview thumbnail
instead of the real one.

This applies to videos a user only has preview access to,
which is a new special case: The event is authorized, but
there is some data (thumbnail, captions, segments and tracks)
that is only viewable when the user has at least read access.