Add segment preview images on text match hover
Also seeing the preview image for the text matches makes it easier to
find a relevant spot in the video.

The segment preview images are not loaded in the initial GQL request as
it would increase response size quite a bit and is not necessary in most
cases. We simply load them on hover, i.e. when they will become used.