Fix z-index issue by removing z-index values
This fix was a easier than anticipated (once I figured this out anyway).
The problem is: giving the `WithTooltip` a `z-index`, means that parent
element (containing both the trigger element and the tooltip), creates
a new stacking context. And a stacking context and its children behaves
as one unit in regards to other stacking contexts. So even setting the
z-index of the tooltip itself to 100, two timelines are two different
stacking contexts, so the 100 of the one tooltip is not rendered in
front of the 4 of the trigger elements of another timeline. So there
were two options:
- Either introduce another div in the floating components that would
  have the trigger element as children, but not the tooltip, and that
  could get a z-index then. But that would have required changing
  appkit.
- Remove the z-index from `WithTooltip` to avoid creating stacking
  contexts. First I thought the only way to do this is to remove the
  big clickable link area of the search results. And we might still do
  that, as it often brings annoying problems. But for this particular
  problem, the solution turned out easier than I thought.