Hide search thumbnails for users without read access
These are supposed to be replaced by a generic preview
image everywhere when a user only has a preview role.
Search events use another table and an api that is missing the
`authorized_data` field, so there needs to be an additional check
to keep the thumbnails hidden.
This poses another issue for authentication which will be
addressed in a later commit.