Add harvest response deserialization tests
This makes sure that we can keep our promise of being able to
deserialize old responses from the Harvest API, as specified by
MIN_REQUIRED_API_VERSION.

This test currently fails, which is the reason I am adding this test
now: because we screwed up. The following commits will fix this.