Fix deserialization of old `Event` items in harvest response
This was an oversight in d8216fa7a32c2594af6ed60a37ecd4b65476afd7
where the `segments` field was added without the `serde(default)`
attribute, therefore making it required.

The two other changes weren't strictly necessary, since `start_time` and
`end_time` are already present in the earliest harvest response format,
but we might want to omit `null` fields instead of explicitly specifying
them as `null` in the future, so this is helps with that.