Move to new branching model
Rename `master` to `main` and add `next` branch for collecting bigger
changes. We ran into the problem of needing to release a bugfix twice
now. Both times, we had changes on `Master` that we did not want to
release with the bug fix. The new goal is to always keep `main` in a
releasable state.