Add endpoints to update acls in tobira (and OC)
This adds a mutation that will send a PUT request to
Opencast to update the acl of a given event.
When the request responds with 204, it is considered
successful and the updated acl is stored in Tobira's DB,
without waiting for sync with OC. That allows us to pretend
that things are happening much faster than they do in actuality,
as Opencast needs to run the "republish metadata" workflow in
order to propagate these changes back to Tobira.

There were some discussions and efforts to change how that is
done in Opencast, but the proposed changes didn't make the cut
yet. Once that or sth similar happens, we should also adapt
the mechanism implemented in this commit.