Add code for debugging/profiling search performance
This shows how long each step in the search "pipeline" takes. This helps
to optimize for faster "search as you type" searches. Typically we don't
commit such debugging code, but this might be useful for a while, to
constantly have a look at the timings. Will very likely be removed some
time in the future.