Remove second 250ms debounce in search
Somehow we did an oopsie and added a second debounce for searching. But
there is already a 30ms debounce! This added lots of delay that the user
notices in a bad way. Of course, less debounce means that the server is
potentially stressed more. But I think this is still absolutely fine.