Lazily render text match timelines
Profiling rendering showed that those match timelines are very expensive
to render. We deferring the rendering, the whole route can be rendered
quicker. A difference of 100ms was quite normal. Also, only timelines
inside the viewport are rendered, which is good for "search as you type"
as most result lists are not viewed by the user at all.