Fix search sometimes ignoring some inputs
Turns out we cancelled the debounced function (that starts the search)
on unmount. I am not sure why we did that. This is just `setTimeout`,
not `setInterval`, so the function is just called at some point and then
the whole thing is over. No need to clean up anything.

The problem with this was that a change in the input field called
`setTimeout(search)`. But this timeout could be cleared by a previous
render-unmount, so that the last letter, for example, might be ignored.