Improve search menu/field handling on mobile
When leaving the search route by clicking on a result, the search mode
is closed, i.e. the normal header is shown. On the search route,
the menu is always in search mode (on narrow screens). We figured that
if users want to get to anything else in the header, they would
intuitively press the arrow-left or the X and get what they want.