Stop merging `read_roles` into `preview_roles`
See https://github.com/elan-ev/tobira/pull/1244#discussion_r1783044285

I thought quite a bit about this and I think both options would have
been fine, but I decided to change this for one reason in particular:
currently, `read_roles` and `write_roles` reflect exactly what Opencast
tells us. We of course also have columns with calculated values, but
we should keep it fairly clear what columns are straight from OC and
which ones are "our columns". Most importantly: if we ever would need
to distinguish between read and preview, this would require a resync.
This would be unfortunate, so I'd rather store the exact OC information
so that we have it, just in case.

I then decided to just adjust the few checks to check both columns, but
we could also have added another column that holds the combination of
the two. That's an implementation detail and we can still change it
later. But yes, the more complicated checks don't seem like a problem
to me.