Remove DB trigger to transfer series credentials to event
From a recent meeting, we concluded that we should just do what the
event ACL tell us. The admins need to make sure the effective (i.e.
after applying merge mode) event ACLs contain the password roles, if
that video should be password protected.

Even if that requirement still changes in the future, I think it should
be done in the harvest code, not as a DB trigger. Further, just always
transferring series credentials to events is also wrong in our more
flexible model (that should cover more use cases than the ETH). Finally,
the trigger as written was not complete as it was missing the "read
roles become preview roles" logic, which is already present in code.

I decided to still store series credentials, as that's useful for a few
reasons, e.g. implementing this feature later on without resyncing.
But this tiny SQL command didn't warrant its own migration, so I
combined both.