Update to `meilisearch_sdk` 0.27.1
Just boring code adjustments, mostly imports. But this is a nice update
as it gets rid of the `isahc` dependency, pruning our dependency tree
a bit.

The `isahc` and `libz-sys` entries in the Cargo.toml were just to
control how indirect dependencies were getting linked and are not
necessary anymore.