Add helper function for mapping ACL arrays to map
The backend serves these ACLs as arrays of objects holding
role, actions and info values. The frontend ACL editor
needs this as a map with role as key and the other two props
as values.

This translation of array to map was already being done
in a few different places and I was about to add another
instance. At that point I think it's ok to just put this
logic in a reusable function.